- Phần thống kê Passed Tests, Failed (viết nhầm chính tả trong phần mềm): mỗi lần làm bài trong từng kĩ năng trên 50% thì coi như Passed (ví dụ luyện Listening 1 lần được từ 12.5 trở lên thì cộng 1 Passed, luyện Reading 1 lần được dưới 17.5 thì cộng 1 Failed).

+++ Utils/ScoresUtils.cs
@@ -15,7 +15,7 @@ namespace TCU.English.Utils
 
         public static float GetThresholdPoint(string typeCode)
         {
-            return (GetMaxScores(typeCode) / (float)2);
+            return (GetMaxScores(typeCode) / 2);
         }
 
         public static float GetMaxScores(string typeCode)
